+++ python/test/integration/test_histogram.py
@@ -28,7 +28,7 @@ def test_count_by_categories():
 
 def test_count_by():
     """Compute histogram with unknown category set"""
-    from opendp.trans import make_split_dataframe, make_select_column, make_resize, make_count_by
+    from opendp.trans import make_split_dataframe, make_select_column, make_resize, make_sized_count_by
     from opendp.meas import make_base_stability
     from opendp.typing import L1Distance, L2Distance
     from opendp.mod import binary_search_chain, enable_features
@@ -44,14 +44,14 @@ def test_count_by():
     laplace_histogram_from_dataframe = binary_search_chain(
         lambda s: (
             preprocess >>
-            make_count_by(size=size, MO=L1Distance[float], TIA=str) >>
+            make_sized_count_by(size=size, MO=L1Distance[float], TIA=str) >>
             make_base_stability(size=size, scale=s, threshold=50., MI=L1Distance[float], TIK=str)
         ), d_in=1, d_out=budget)
 
     gaussian_histogram_from_dataframe = binary_search_chain(
         lambda s: (
             preprocess >>
-            make_count_by(size=size, MO=L2Distance[float], TIA=str) >>
+            make_sized_count_by(size=size, MO=L2Distance[float], TIA=str) >>
             make_base_stability(size=size, scale=s, threshold=50., MI=L2Distance[float], TIK=str)
         ), d_in=1, d_out=budget)