+++ rust/opendp-ffi/src/trans/count.rs
@@ -2,14 +2,14 @@ use std::convert::TryFrom;
 use std::hash::Hash;
 use std::os::raw::{c_char, c_uint};
 
-use num::{Bounded, Integer, One, Zero};
+use num::{Bounded, Integer, One, Zero, Float};
 use num::traits::FloatConst;
 
 use opendp::core::{SensitivityMetric};
 use opendp::dist::{L1Distance, L2Distance, IntDistance};
 use opendp::err;
 use opendp::traits::{DistanceConstant, InfCast, ExactIntCast, SaturatingAdd, CheckNull};
-use opendp::trans::{CountByConstant, make_count, make_count_by, make_count_by_categories, make_count_distinct};
+use opendp::trans::{CountByConstant, make_count, make_sized_count_by, make_count_by_categories, make_count_distinct, make_sized_count_by_categories, SizedCountByConstant};
 
 use crate::any::{AnyObject, AnyTransformation};
 use crate::any::Downcast;
@@ -59,41 +59,85 @@ pub extern "C" fn opendp_trans__make_count_distinct(
 #[no_mangle]
 pub extern "C" fn opendp_trans__make_count_by_categories(
     categories: *const AnyObject,
-    MO: *const c_char, TI: *const c_char, TO: *const c_char,
+    MO: *const c_char, TIA: *const c_char
 ) -> FfiResult<*mut AnyTransformation> {
+
     fn monomorphize<QO>(
         categories: *const AnyObject,
-        MO: Type, TI: Type, TO: Type,
+        MO: Type, TIA: Type
     ) -> FfiResult<*mut AnyTransformation>
-        where QO: DistanceConstant<IntDistance> + One,
+        where QO: DistanceConstant<IntDistance> + One + Integer + Zero + SaturatingAdd + CheckNull,
               IntDistance: InfCast<QO> {
-        fn monomorphize2<MO, TI, TO>(categories: *const AnyObject) -> FfiResult<*mut AnyTransformation>
+
+        fn monomorphize2<MO, TIA>(categories: *const AnyObject) -> FfiResult<*mut AnyTransformation>
             where MO: 'static + SensitivityMetric + CountByConstant<MO::Distance>,
-                  MO::Distance: DistanceConstant<IntDistance> + One,
-                  TI: 'static + Eq + Hash + Clone + CheckNull,
-                  TO: 'static + Integer + Zero + One + SaturatingAdd + CheckNull,
+                  MO::Distance: DistanceConstant<IntDistance> + One + Integer + Zero + SaturatingAdd + CheckNull,
+                  TIA: 'static + Eq + Hash + Clone + CheckNull,
                   IntDistance: InfCast<MO::Distance> {
-            let categories = try_!(try_as_ref!(categories).downcast_ref::<Vec<TI>>()).clone();
-            make_count_by_categories::<MO, TI, TO>(categories).into_any()
+
+            let categories = try_!(try_as_ref!(categories).downcast_ref::<Vec<TIA >>()).clone();
+            make_count_by_categories::<MO, TIA>(categories).into_any()
         }
         dispatch!(monomorphize2, [
             (MO, [L1Distance<QO>, L2Distance<QO>]),
-            (TI, @hashable),
-            (TO, @integers)
+            (TIA, @hashable)
         ], (categories))
     }
     let MO = try_!(Type::try_from(MO));
-    let TI = try_!(Type::try_from(TI));
-    let TO = try_!(Type::try_from(TO));
+    let TIA = try_!(Type::try_from(TIA));
 
     let QO = try_!(MO.get_sensitivity_distance());
     dispatch!(monomorphize, [
         (QO, @integers)
-    ], (categories, MO, TI, TO))
+    ], (categories, MO, TIA))
+}
+
+
+#[no_mangle]
+pub extern "C" fn opendp_trans__make_sized_count_by_categories(
+    size: c_uint, categories: *const AnyObject,
+    MO: *const c_char, TIA: *const c_char, TOA: *const c_char,
+) -> FfiResult<*mut AnyTransformation> {
+
+    fn monomorphize<QO>(
+        size: usize, categories: *const AnyObject,
+        MO: Type, TIA: Type, TOA: Type,
+    ) -> FfiResult<*mut AnyTransformation>
+        where QO: DistanceConstant<IntDistance> + One + Float + InfCast<u8>,
+              IntDistance: InfCast<QO> {
+
+        fn monomorphize2<MO, TIA, TOA>(
+            size: usize, categories: *const AnyObject
+        ) -> FfiResult<*mut AnyTransformation>
+            where MO: 'static + SensitivityMetric + SizedCountByConstant<MO::Distance>,
+                  MO::Distance: DistanceConstant<IntDistance> + One + Float + InfCast<u8>,
+                  TIA: 'static + Eq + Hash + Clone + CheckNull,
+                  TOA: 'static + Integer + Zero + One + SaturatingAdd + CheckNull,
+                  IntDistance: InfCast<MO::Distance> {
+
+            let categories = try_!(try_as_ref!(categories).downcast_ref::<Vec<TIA >>()).clone();
+            make_sized_count_by_categories::<MO, TIA, TOA>(size, categories).into_any()
+        }
+
+        dispatch!(monomorphize2, [
+            (MO, [L1Distance<QO>, L2Distance<QO>]),
+            (TIA, @hashable),
+            (TOA, @integers)
+        ], (size, categories))
+    }
+    let size = size as usize;
+    let MO = try_!(Type::try_from(MO));
+    let TIA = try_!(Type::try_from(TIA));
+    let TOA = try_!(Type::try_from(TOA));
+
+    let QO = try_!(MO.get_sensitivity_distance());
+    dispatch!(monomorphize, [
+        (QO, @floats)
+    ], (size, categories, MO, TIA, TOA))
 }
 
 #[no_mangle]
-pub extern "C" fn opendp_trans__make_count_by(
+pub extern "C" fn opendp_trans__make_sized_count_by(
     size: c_uint,
     MO: *const c_char, TIA: *const c_char, TOA: *const c_char,
 ) -> FfiResult<*mut AnyTransformation> {
@@ -108,7 +152,7 @@ pub extern "C" fn opendp_trans__make_count_by(
                   TIA: 'static + Eq + Hash + Clone + CheckNull,
                   TOA: 'static + Integer + Zero + One + SaturatingAdd + CheckNull,
                   IntDistance: InfCast<MO::Distance> {
-            make_count_by::<MO, TIA, TOA>(size).into_any()
+            make_sized_count_by::<MO, TIA, TOA>(size).into_any()
         }
         dispatch!(monomorphize2, [
             (MO, [L1Distance<QO>, L2Distance<QO>]),